create a separate file for scene testing
